+++ src/OpenstoreSchema/Core/Tools/Console/Command/Fixture/LoadFixtureCommand.php
@@ -1,5 +1,15 @@
 <?php
 
+/*
+ * opentore-schema-core
+ *
+ * @author    Vanvelthem Sébastien
+ * @link      https://github.com/belgattitude/openstore-schema-core
+ * @copyright Copyright (c) 2015-2017 Vanvelthem Sébastien
+ * @license   MIT License, https://github.com/belgattitude/openstore-schema-core/blob/master/LICENSE.md
+ *
+ */
+
 namespace OpenstoreSchema\Core\Tools\Console\Command\Fixture;
 
 use Symfony\Component\Console\Input\InputOption;